API Refactoring
