+++ src/DataFixtures/DevopsFixture.php
@@ -2,8 +2,8 @@
 
 namespace App\DataFixtures;
 
-use App\Entity\Internal\Edition;
-use App\Entity\Internal\Talk;
+use App\Entity\Edition;
+use App\Entity\Talk;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 use Mni\FrontYAML\Parser;
@@ -57,7 +57,7 @@ class DevopsFixture extends Fixture implements DependentFixtureInterface
                 $talk->setTitle($subject['title']);
                 $talk->setType($this->getReference(AppFixtures::TT_REGULAR));
                 foreach ($subject['speakers'] as $slug) {
-                    $speaker= $manager->getRepository('App\Entity\Internal\Speaker')->findOneBySlug($slug);
+                    $speaker= $manager->getRepository('App\Entity\Speaker')->findOneBySlug($slug);
                     if (!!$speaker) {
                         $talk->addSpeaker($speaker);
                     }
