+++ src/Entity/Edition.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Entity\Internal;
+namespace App\Entity;
 
 use ApiPlatform\Core\Annotation\ApiResource;
-use App\Repository\Internal\EditionRepository;
+use App\Repository\EditionRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
@@ -11,8 +11,8 @@ use Gedmo\Mapping\Annotation as Gedmo;
 
 /**
  * @ORM\Entity(repositoryClass=EditionRepository::class)
- * @ApiResource()
  */
+#[ApiResource]
 class Edition
 {
     /**
