--- src/Entity/Internal/Place.php
@@ -1,273 +1,0 @@
-<?php
-
-namespace App\Entity\Internal;
-
-use App\Repository\Internal\PlaceRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Mapping\Annotation as Gedmo;
-
-/**
- * @ORM\Entity(repositoryClass=PlaceRepository::class)
- */
-class Place
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $name;
-
-    /**
-     * @Gedmo\Slug(fields={"name"})
-     * @ORM\Column(length=255,unique=true)
-     */
-    private $slug;
-
-    /**
-     * @ORM\Column(type="text", nullable=true)
-     */
-    private $description;
-
-    /**
-     * @ORM\Column(type="string", length=500, nullable=true)
-     */
-    private $url;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $address1;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $address2;
-
-    /**
-     * @ORM\Column(type="string", length=10)
-     */
-    private $postal_code;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $city;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $coutry;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $logo;
-
-    /**
-     * @ORM\OneToMany(targetEntity=Edition::class, mappedBy="Place")
-     */
-    private $editions;
-
-    /**
-     * @ORM\Column(type="text", nullable=true)
-     */
-    private $description_html;
-
-    /**
-     * @ORM\Column(type="text", nullable=true)
-     */
-    private $description_markdown;
-
-    public function __construct()
-    {
-        $this->editions = new ArrayCollection();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getSlug()
-    {
-        return $this->slug;
-    }
-
-
-    public function setSlug(string $slug): self
-    {
-        $this->slug = $slug;
-
-        return $this;
-    }
-
-    public function getDescription(): ?string
-    {
-        return $this->description;
-    }
-
-    public function setDescription(?string $description): self
-    {
-        $this->description = $description;
-
-        return $this;
-    }
-
-    public function getUrl(): ?string
-    {
-        return $this->url;
-    }
-
-    public function setUrl(?string $url): self
-    {
-        $this->url = $url;
-
-        return $this;
-    }
-
-    public function getAddress1(): ?string
-    {
-        return $this->address1;
-    }
-
-    public function setAddress1(string $address1): self
-    {
-        $this->address1 = $address1;
-
-        return $this;
-    }
-
-    public function getAddress2(): ?string
-    {
-        return $this->address2;
-    }
-
-    public function setAddress2(?string $address2): self
-    {
-        $this->address2 = $address2;
-
-        return $this;
-    }
-
-    public function getPostalCode(): ?string
-    {
-        return $this->postal_code;
-    }
-
-    public function setPostalCode(string $postal_code): self
-    {
-        $this->postal_code = $postal_code;
-
-        return $this;
-    }
-
-    public function getCity(): ?string
-    {
-        return $this->city;
-    }
-
-    public function setCity(string $city): self
-    {
-        $this->city = $city;
-
-        return $this;
-    }
-
-    public function getCoutry(): ?string
-    {
-        return $this->coutry;
-    }
-
-    public function setCoutry(?string $coutry): self
-    {
-        $this->coutry = $coutry;
-
-        return $this;
-    }
-
-    public function getLogo(): ?string
-    {
-        return $this->logo;
-    }
-
-    public function setLogo(?string $logo): self
-    {
-        $this->logo = $logo;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Edition[]
-     */
-    public function getEditions(): Collection
-    {
-        return $this->editions;
-    }
-
-    public function addEdition(Edition $edition): self
-    {
-        if (!$this->editions->contains($edition)) {
-            $this->editions[] = $edition;
-            $edition->setPlace($this);
-        }
-
-        return $this;
-    }
-
-    public function removeEdition(Edition $edition): self
-    {
-        if ($this->editions->removeElement($edition)) {
-            // set the owning side to null (unless already changed)
-            if ($edition->getPlace() === $this) {
-                $edition->setPlace(null);
-            }
-        }
-
-        return $this;
-    }
-
-    public function getDescriptionHtml(): ?string
-    {
-        return $this->description_html;
-    }
-
-    public function setDescriptionHtml(?string $description_html): self
-    {
-        $this->description_html = $description_html;
-
-        return $this;
-    }
-
-    public function getDescriptionMarkdown(): ?string
-    {
-        return $this->description_markdown;
-    }
-
-    public function setDescriptionMarkdown(?string $description_markdown): self
-    {
-        $this->description_markdown = $description_markdown;
-
-        return $this;
-    }
-}
