+++ src/Entity/schemaOrg/CreativeWork.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Entity;
+namespace App\Entity\schemaOrg;
 
 use ApiPlatform\Core\Annotation\ApiProperty;
 use ApiPlatform\Core\Annotation\ApiResource;
@@ -13,7 +13,7 @@ use Symfony\Component\Validator\Constraints as Assert;
  *
  * @see http://schema.org/CreativeWork Documentation on Schema.org
  *
- * @ApiResource(
+ * ApiResource(
  *      iri="http://schema.org/CreativeWork",
  *      collectionOperations={},
  *      itemOperations={"get"}
