+++ src/Entity/schemaOrg/Place.php
@@ -1,0 +1,131 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Entity\schemaOrg;
+
+use ApiPlatform\Core\Annotation\ApiProperty;
+use ApiPlatform\Core\Annotation\ApiResource;
+use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Serializer\Annotation\Groups;
+
+/**
+ * Entities that have a somewhat fixed, physical extension.
+ *
+ * @see http://schema.org/Place Documentation on Schema.org
+ *
+ * ApiResource(iri="http://schema.org/Place",
+ *      collectionOperations={"get"},
+ *      itemOperations={"get"},
+ *      normalizationContext={"groups"={"place"}},
+ *      denormalizationContext={"groups"={"place"}}
+ * )
+ */
+class Place
+{
+    /**
+     * @var string|null
+     * @ApiProperty(identifier=true)
+     */
+    private $id;
+
+    /**
+     * @var string|null the name of the item
+     *
+     * @ApiProperty(iri="http://schema.org/name")
+     * @Groups({"event", "place"})
+     */
+    private $name;
+
+    /**
+     * @var string|null a description of the item
+     *
+     * @ApiProperty(iri="http://schema.org/description")
+     * @Groups({"event", "place"})
+     */
+    private $description;
+
+    /**
+     * @var PostalAddress|null physical address of the item
+     *
+     * @ApiProperty(iri="http://schema.org/address")
+     * @Groups({"event", "place"})
+     */
+    private $address;
+
+    /**
+     * @var string|null The identifier property represents any kind of identifier for any kind of \[\[Thing\]\], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See \[background notes\](/docs/datamodel.html#identifierBg) for more details.
+     *
+     * @ApiProperty(iri="http://schema.org/identifier")
+     * @Assert\Url
+     */
+    private $identifier;
+
+    /**
+     * @var string|null An image of the item. This can be a \[\[URL\]\] or a fully described \[\[ImageObject\]\].
+     *
+     * @ApiProperty(iri="http://schema.org/image")
+     * @Assert\Url
+     */
+    private $image;
+
+    public function setId(string $id): void
+    {
+        $this->id = $id;
+    }
+
+    public function getId(): ?string
+    {
+        return $this->id;
+    }
+
+    public function setName(?string $name): void
+    {
+        $this->name = $name;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setDescription(?string $description): void
+    {
+        $this->description = $description;
+    }
+
+    public function getDescription(): ?string
+    {
+        return $this->description;
+    }
+
+    public function setAddress(?PostalAddress $address): void
+    {
+        $this->address = $address;
+    }
+
+    public function getAddress(): ?PostalAddress
+    {
+        return $this->address;
+    }
+
+    /* public function setIdentifier(?string $identifier): void */
+    /* { */
+    /*     $this->identifier = $identifier; */
+    /* } */
+
+    /* public function getIdentifier(): ?string */
+    /* { */
+    /*     return $this->identifier; */
+    /* } */
+
+    public function setImage(?string $image): void
+    {
+        $this->image = $image;
+    }
+
+    public function getImage(): ?string
+    {
+        return $this->image;
+    }
+}
